FEXCore: Move EnumOperators to FEXCore
fextl needs this and can't depend on FHU
