+++ External/FEXCore/include/FEXCore/Utils/EnumOperators.h
